refactor: move manager install functionalities to a separate file
