+++ ext/exd/manager.go
@@ -2,11 +2,9 @@ package exd
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"path"
-	"time"
 )
 
 // ExtensionDir is directory path where to store the extensions
@@ -31,193 +29,6 @@ func NewManager(ctx context.Context, httpDoer HTTPDoer, manifester Manifester, i
 		manifester: manifester,
 		installer:  installer,
 	}, nil
-}
-
-// Install installs extension based on the remote path
-func (m *Manager) Install(remotePath, commandName string) error {
-	if err := m.validateRemotePath(remotePath); err != nil {
-		return formatError("error validating installation for [%s]: %w", remotePath, err)
-	}
-
-	manifest, err := m.manifester.Load(ExtensionDir)
-	if err != nil {
-		return formatError("error loading manifest: %w", err)
-	}
-
-	remoteMetadata, err := m.extractMetadata(remotePath)
-	if err != nil {
-		return formatError("error extracting remote metadata for: %w", err)
-	}
-
-	client, err := m.findClientProvider(remoteMetadata.ProviderName)
-	if err != nil {
-		return formatError("error finding client provider [%s]: %w", remoteMetadata.ProviderName, err)
-	}
-
-	release, err := client.GetRelease(remoteMetadata.AssetAPIPath)
-	if err != nil {
-		return formatError("error getting release for [%s/%s@%s]: %w",
-			remoteMetadata.OwnerName, remoteMetadata.RepoName, remoteMetadata.TagName, err,
-		)
-	}
-	m.updateRemoteMetadata(remoteMetadata, release, commandName)
-
-	if err := m.validateRemoteMetadataToManifest(remoteMetadata, manifest); err != nil {
-		return formatError("error remote metadata on manifest: %w", err)
-	}
-
-	if m.isAlreadyInstalled(manifest, remoteMetadata) {
-		return formatError("[%s/%s@%s] is already installed",
-			remoteMetadata.OwnerName, remoteMetadata.RepoName, remoteMetadata.TagName,
-		)
-	}
-
-	asset, err := client.DownloadAsset(remoteMetadata.AssetAPIPath)
-	if err != nil {
-		return formatError("error downloading asset for [%s/%s@%s]: %w",
-			remoteMetadata.OwnerName, remoteMetadata.RepoName, remoteMetadata.TagName, err,
-		)
-	}
-
-	if err := m.installAsset(asset, remoteMetadata); err != nil {
-		return formatError("error installing asset for [%s/%s@%s]: %w",
-			remoteMetadata.OwnerName, remoteMetadata.RepoName, remoteMetadata.TagName, err,
-		)
-	}
-
-	m.updateManifest(manifest, remoteMetadata, release)
-	if err := m.manifester.Flush(manifest, ExtensionDir); err != nil {
-		return formatError("error flushing manifest: %w", err)
-	}
-	return nil
-}
-
-func (m *Manager) updateManifest(manifest *Manifest, remoteMetadata *RemoteMetadata, release *RepositoryRelease) {
-	if updated := m.updateExistingProjectInManifest(manifest, remoteMetadata, release); !updated {
-		m.addNewProjectToManifest(manifest, remoteMetadata, release)
-	}
-	manifest.UpdatedAt = time.Now()
-}
-
-func (*Manager) addNewProjectToManifest(manifest *Manifest, remoteMetadata *RemoteMetadata, release *RepositoryRelease) {
-	manifest.RepositoryOwners = append(manifest.RepositoryOwners, &RepositoryOwner{
-		Name:     remoteMetadata.OwnerName,
-		Provider: remoteMetadata.ProviderName,
-		Projects: []*RepositoryProject{
-			{
-				Name:          remoteMetadata.RepoName,
-				CommandName:   remoteMetadata.CommandName,
-				ActiveTagName: remoteMetadata.TagName,
-				AssetAPIPath:  remoteMetadata.AssetAPIPath,
-				AssetDirPath:  remoteMetadata.AssetDirPath,
-				Releases:      []*RepositoryRelease{release},
-			},
-		},
-	})
-}
-
-func (*Manager) updateExistingProjectInManifest(manifest *Manifest, remoteMetadata *RemoteMetadata, release *RepositoryRelease) bool {
-	for _, owner := range manifest.RepositoryOwners {
-		if owner.Name == remoteMetadata.OwnerName {
-			for _, project := range owner.Projects {
-				if project.Name == remoteMetadata.RepoName {
-					project.ActiveTagName = remoteMetadata.TagName
-					project.AssetAPIPath = remoteMetadata.AssetAPIPath
-					project.AssetDirPath = remoteMetadata.AssetDirPath
-					project.CommandName = remoteMetadata.CommandName
-					project.Releases = append(project.Releases, release)
-					return true
-				}
-			}
-			break
-		}
-	}
-	return false
-}
-
-func (m *Manager) installAsset(asset []byte, remoteMetadata *RemoteMetadata) error {
-	if err := m.installer.Prepare(remoteMetadata); err != nil {
-		return fmt.Errorf("error preparing installation: %w", err)
-	}
-	return m.installer.Install(asset, remoteMetadata)
-}
-
-func (*Manager) validateRemoteMetadataToManifest(remoteMetadata *RemoteMetadata, manifest *Manifest) error {
-	for _, owner := range manifest.RepositoryOwners {
-		for _, project := range owner.Projects {
-			if project.CommandName == remoteMetadata.CommandName {
-				if owner.Name == remoteMetadata.OwnerName && project.Name == remoteMetadata.RepoName {
-					return nil
-				}
-				return fmt.Errorf("command [%s] is already used by [%s/%s@%s]",
-					remoteMetadata.CommandName, owner.Name, project.Name, project.ActiveTagName,
-				)
-			}
-		}
-	}
-	return nil
-}
-
-func (*Manager) isAlreadyInstalled(manifest *Manifest, remoteMetadata *RemoteMetadata) bool {
-	for _, owner := range manifest.RepositoryOwners {
-		if owner.Name == remoteMetadata.OwnerName {
-			for _, project := range owner.Projects {
-				if project.Name == remoteMetadata.RepoName {
-					for _, release := range project.Releases {
-						if release.TagName == remoteMetadata.TagName {
-							return true
-						}
-					}
-					return false
-				}
-			}
-			return false
-		}
-	}
-	return false
-}
-
-func (*Manager) updateRemoteMetadata(remoteMetadata *RemoteMetadata, release *RepositoryRelease, commandName string) {
-	remoteMetadata.TagName = release.TagName
-	if commandName != "" {
-		remoteMetadata.CommandName = commandName
-	}
-}
-
-func (m *Manager) findClientProvider(provider string) (Client, error) {
-	newClient, err := NewClientRegistry.Get(provider)
-	if err != nil {
-		return nil, fmt.Errorf("error getting new client: %w", err)
-	}
-	return newClient(m.ctx, m.httpDoer)
-}
-
-func (*Manager) extractMetadata(remotePath string) (*RemoteMetadata, error) {
-	var remoteMetadata *RemoteMetadata
-	for _, parseFn := range ParseRegistry {
-		mtdt, err := parseFn(remotePath)
-		if errors.Is(err, ErrUnrecognizedRemotePath) {
-			continue
-		}
-		if err != nil {
-			return nil, fmt.Errorf("errors parsing [%s]: %w", remotePath, err)
-		}
-		if mtdt != nil {
-			remoteMetadata = mtdt
-			break
-		}
-	}
-	if remoteMetadata == nil {
-		return nil, fmt.Errorf("[%s] is not recognized", remotePath)
-	}
-	return remoteMetadata, nil
-}
-
-func (m *Manager) validateRemotePath(remotePath string) error {
-	if remotePath == "" {
-		return ErrEmptyRemotePath
-	}
-	return validate(m.ctx, m.httpDoer, m.manifester, m.installer)
 }
 
 func validate(ctx context.Context, httpDoer HTTPDoer, manifester Manifester, installer Installer) error {
