+++ ext/exd/manager_activate_test.go
@@ -1,0 +1,1 @@
+package exd_test
